2.1
Still sweeping to remove strange formatting and animations. Label system is getting frustrating to navigate. I am used to working with CSS section by section, a lot of this stuff is linked globally, but that is creating issues like gray font everywhere (even in places where it should be white for accessibility)